[Decomp][Player] Cohesion split: extract talent system to PlayerTalent.cpp
Re-applies the mangosthree decomposition onto MangosOne's own TBC 2.4.3
source. Pure code move out of the parent translation unit into the new
cohesively-named file(s); no behaviour change (source diff is deletion-only).
CMake file(GLOB) picks the new TU(s) up automatically. Verified: game.lib
compiles and links clean (MSVC, Release).

Not ported (absent in MangosOne TBC 2.4.3 -- these are
later-expansion methods that do not exist in this branch):
  Player::LearnPetTalent,Player::BuildPlayerTalentsInfoData,Player::BuildPetTalentsInfoData,Player::SendTalentsInfoData,Player::BuildEnchantmentsInfoData,Player::SendEquipmentSetList,Player::SetEquipmentSet,Player::DeleteEquipmentSet,Player::ActivateSpec,Player::UpdateSpecCount
Player::LearnPetTalent,Player::BuildPlayerTalentsInfoData,Player::BuildPetTalentsInfoData,Player::SendTalentsInfoData,Player::BuildEnchantmentsInfoData,Player::SendEquipmentSetList,Player::SetEquipmentSet,Player::DeleteEquipmentSet,Player::ActivateSpec,Player::UpdateSpecCount
Ported-from: mangosthree/server@8e9e5122bd54
