+++ src/game/Object/Player.cpp
@@ -12654,173 +12654,6 @@ void Player::HandleFall(MovementInfo const& movementInfo)
 }
 
 /**
- * @brief Learns a specific talent rank if all requirements are satisfied.
- *
- * @param talentId The talent entry identifier.
- * @param talentRank The requested talent rank index.
- */
-void Player::LearnTalent(uint32 talentId, uint32 talentRank)
-{
-    uint32 CurTalentPoints = GetFreeTalentPoints();
-
-    if (CurTalentPoints == 0)
-    {
-        return;
-    }
-
-    if (talentRank >= MAX_TALENT_RANK)
-    {
-        return;
-    }
-
-    TalentEntry const* talentInfo = sTalentStore.LookupEntry(talentId);
-
-    if (!talentInfo)
-    {
-        return;
-    }
-
-    TalentTabEntry const* talentTabInfo = sTalentTabStore.LookupEntry(talentInfo->TalentTab);
-
-    if (!talentTabInfo)
-    {
-        return;
-    }
-
-    // prevent learn talent for different class (cheating)
-    if ((getClassMask() & talentTabInfo->ClassMask) == 0)
-    {
-        return;
-    }
-
-    // find current max talent rank
-    uint32 curtalent_maxrank = 0;
-    for (int32 k = MAX_TALENT_RANK - 1; k > -1; --k)
-    {
-        if (talentInfo->RankID[k] && HasSpell(talentInfo->RankID[k]))
-        {
-            curtalent_maxrank = k + 1;
-            break;
-        }
-    }
-
-    // we already have same or higher talent rank learned
-    if (curtalent_maxrank >= (talentRank + 1))
-    {
-        return;
-    }
-
-    // check if we have enough talent points
-    if (CurTalentPoints < (talentRank - curtalent_maxrank + 1))
-    {
-        return;
-    }
-
-    // Check if it requires another talent
-    if (talentInfo->DependsOn > 0)
-    {
-        if (TalentEntry const* depTalentInfo = sTalentStore.LookupEntry(talentInfo->DependsOn))
-        {
-            bool hasEnoughRank = false;
-            for (int i = talentInfo->DependsOnRank; i < MAX_TALENT_RANK; ++i)
-            {
-                if (depTalentInfo->RankID[i] != 0)
-                    if (HasSpell(depTalentInfo->RankID[i]))
-                    {
-                        hasEnoughRank = true;
-                    }
-            }
-
-            if (!hasEnoughRank)
-            {
-                return;
-            }
-        }
-    }
-
-    // Check if it requires spell
-    if (talentInfo->DependsOnSpell && !HasSpell(talentInfo->DependsOnSpell))
-    {
-        return;
-    }
-
-    // Find out how many points we have in this field
-    uint32 spentPoints = 0;
-
-    uint32 tTab = talentInfo->TalentTab;
-    if (talentInfo->Row > 0)
-    {
-        unsigned int numRows = sTalentStore.GetNumRows();
-        for (unsigned int i = 0; i < numRows; ++i)          // Loop through all talents.
-        {
-            // Someday, someone needs to revamp
-            const TalentEntry* tmpTalent = sTalentStore.LookupEntry(i);
-            if (tmpTalent)                                  // the way talents are tracked
-            {
-                if (tmpTalent->TalentTab == tTab)
-                {
-                    for (int j = 0; j < MAX_TALENT_RANK; ++j)
-                    {
-                        if (tmpTalent->RankID[j] != 0)
-                        {
-                            if (HasSpell(tmpTalent->RankID[j]))
-                            {
-                                spentPoints += j + 1;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // not have required min points spent in talent tree
-    if (spentPoints < (talentInfo->Row * MAX_TALENT_RANK))
-    {
-        return;
-    }
-
-    // spell not set in talent.dbc
-    uint32 spellid = talentInfo->RankID[talentRank];
-    if (spellid == 0)
-    {
-        sLog.outError("Talent.dbc have for talent: %u Rank: %u spell id = 0", talentId, talentRank);
-        return;
-    }
-
-    // already known
-    if (HasSpell(spellid))
-    {
-        return;
-    }
-
-    // learn! (other talent ranks will unlearned at learning)
-    learnSpell(spellid, false);
-    DETAIL_LOG("TalentID: %u Rank: %u Spell: %u\n", talentId, talentRank, spellid);
-#ifdef ENABLE_ELUNA
-    if (Eluna* e = GetEluna())
-    {
-        e->OnLearnTalents(this, talentId, talentRank, spellid);
-    }
-#endif /*ENABLE_ELUNA*/
-
-}
-
-/**
- * @brief Refreshes stored fall tracking data when movement indicates a new fall state.
- *
- * @param minfo The current movement information.
- * @param opcode The movement opcode being processed.
- */
-void Player::UpdateFallInformationIfNeed(MovementInfo const& minfo, uint16 opcode)
-{
-    if (m_lastFallTime >= minfo.GetFallTime() || m_lastFallZ <= minfo.GetPos()->z || opcode == MSG_MOVE_FALL_LAND)
-    {
-        SetFallInformation(minfo.GetFallTime(), minfo.GetPos()->z);
-    }
-}
-
-/**
  * @brief Temporarily unsummons the current pet when the player's state requires it.
  */
 void Player::UnsummonPetTemporaryIfAny()
@@ -12867,31 +12700,6 @@ void Player::ResummonPetTemporaryUnSummonedIfAny()
     }
 
     m_temporaryUnsummonedPetNumber = 0;
-}
-
-/**
- * @brief Checks whether the player can currently see spell-click interaction on a creature.
- *
- * @param c The creature to evaluate.
- * @return True if spell-click should be visible; otherwise, false.
- */
-bool Player::canSeeSpellClickOn(Creature const* c) const
-{
-    if (!c->HasFlag(UNIT_NPC_FLAGS, UNIT_NPC_FLAG_SPELLCLICK))
-    {
-        return false;
-    }
-
-    SpellClickInfoMapBounds clickPair = sObjectMgr.GetSpellClickInfoMapBounds(c->GetEntry());
-    for (SpellClickInfoMap::const_iterator itr = clickPair.first; itr != clickPair.second; ++itr)
-    {
-        if (itr->second.IsFitToRequirements(this, c))
-        {
-            return true;
-        }
-    }
-
-    return false;
 }
 
 /**
